Refactor 3rd party package structure
